requirements.txt
python-telegram-bot==13.7
schedule==1.1.0
python-dotenv==0.19.0
